% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_knn.R
\name{bh_knn}
\alias{bh_knn}
\title{Run KNN}
\usage{
bh_knn(train, test, labels, k = 1, dist = function(x, y) {     abs(cor(x,
  y)) })
}
\arguments{
\item{train}{matrix or data frame of training set cases}

\item{test}{matrix or data frame of test set cases}

\item{labels}{multiclass factor of true classifcations}

\item{k}{number of neighbours considered}

\item{pred}{the class to predict}
}
\description{
Run KNN
}

