% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_knn_sign.R
\name{bh_knn_sign}
\alias{bh_knn_sign}
\title{Run KNN}
\usage{
bh_knn_sign(train, test, labels, k = 1)
}
\arguments{
\item{train}{matrix or data frame of training set cases}

\item{test}{matrix or data frame of test set cases}

\item{labels}{multiclass factor of true classifcations}
}
\description{
Run KNN
}

