% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_loocv.R
\name{bh_loocv}
\alias{bh_loocv}
\title{Run Leave One Out Cross-Validation}
\usage{
bh_loocv(classifier, train, labels)
}
\arguments{
\item{train}{matrix or data frame of training set cases}

\item{labels}{multiclass factor of true classifcations}

\item{classifer}{the classifier function}
}
\description{
Run Leave One Out Cross-Validation
}

