% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_plot_auc.R
\name{bh_plot_auc}
\alias{bh_plot_auc}
\title{Run Leave One Out Cross-Validation}
\usage{
bh_plot_auc(target_pred, target_class)
}
\arguments{
\item{target_pred}{the binary predictions}

\item{target_class}{the labeled data}
}
\description{
Run Leave One Out Cross-Validation
}

